requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.36
diffusers>=0.25
opencv-python>=4.8
numpy>=1.23
Pillow>=10.0
tqdm>=4.66
torch-fidelity>=0.3
